+++ js/util.js
@@ -1,0 +1,15 @@
+const getRandomPositiveInteger = (min, max) => {
+  if (min < 0 || max < 0) {
+    return NaN;
+  }
+  const lower = Math.ceil(Math.min(min, max));
+  const upper = Math.floor(Math.max(min, max));
+
+  return Math.floor(Math.random() * (upper - lower + 1) + lower);
+};
+
+const checkCommentLength = (newComment, maxLength) => newComment.length <= maxLength;
+
+const getRandomArrayElement = (elements) => elements[getRandomPositiveInteger(0, elements.length - 1)];
+
+export {getRandomPositiveInteger, getRandomArrayElement, checkCommentLength};
